Refactor inspector
