+++ dwds/lib/src/utilities/shared.dart
@@ -4,8 +4,9 @@
 
 import 'dart:io';
 
-import 'package:dwds/dwds.dart' show ModuleStrategy;
+import 'package:webkit_inspection_protocol/webkit_inspection_protocol.dart';
 
+import '../../dwds.dart' show ChromeDebugException, ModuleStrategy;
 import '../utilities/wrapped_service.dart';
 
 VMRef toVMRef(VM vm) => VMRef()..name = vm.name;
@@ -56,4 +57,33 @@ String get getLibraries {
   }
   return expression +=
       "let libs = $loadModule('dart_sdk').dart.getLibraries();\n";
+}
+
+/// Creates a snippet of JS code that initializes a `library` variable that has
+/// the actual library object in DDC for [libraryUri].
+///
+/// In DDC we have module libraries indexed by names of the form
+/// 'packages/package/mainFile' with no .dart suffix on the file, or
+/// 'directory/packageName/mainFile', also with no .dart suffix, and relative to
+/// the serving root, normally /web within the package. These modules have a map
+/// from the URI with a Dart-specific scheme (package: or org-dartlang-app:) to
+/// the library objects. The [libraryUri] parameter should be one of these
+/// Dart-specific scheme URIs, and we set `library` the corresponding library.
+String getLibrarySnippet(String libraryUri) => '''
+   var sdkUtils = $loadModule('dart_sdk').dart;
+   var library = sdkUtils.getLibrary('$libraryUri');
+   if (!library) throw 'cannot find library for $libraryUri';
+  ''';
+
+/// Throws an [ExceptionDetails] object if `exceptionDetails` is present on the
+/// result.
+void handleErrorIfPresent(WipResponse response,
+    {String evalContents, Object additionalDetails}) {
+  if (response == null) return;
+  if (response.result.containsKey('exceptionDetails')) {
+    throw ChromeDebugException(
+        response.result['exceptionDetails'] as Map<String, dynamic>,
+        evalContents: evalContents,
+        additionalDetails: additionalDetails);
+  }
 }
