+++ dwds/test/fixtures/fakes.dart
@@ -69,7 +69,7 @@ class FakeInspector extends Domain implements AppInspector {
   @override
   IsolateRef get isolateRef => null;
   @override
-  InstanceHelper get instanceHelper => InstanceHelper(null, null, null);
+  InstanceHelper get instanceHelper => InstanceHelper(null);
 }
 
 class FakeSseConnection implements SseConnection {
